Microservicios/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.89.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
cryptography==3.4.8
python-jose==3.3.0  # Para manejar tokens JWT con 'jose'

